Add scaffolding.
We’ll store package-agnostic docs in /content. We’ll use docs-fetcher
to fetch our docs from the package repos. We’ll use docs-preprocessor
to convert package folder structures to one better suited to a docs
page. We’ll use docs-generator to generate HTML from markdown. We’ll
use docs.yml to declare which packages/repos to fetch and the ordering
of folders in content.