Add a new support page (#86)
* Add a new support page

In addition to adding the content for the page itself, this change
removes the link from the Guides section sidebar, but adds it to
the footer. (It's also already linked to from the header nav.)

New Card and Grid components have been added to support richer
on .mdx pages.

This commit also removes the unused zero2hero content, which
included some mdx files which we don't need to process. This
content will be added later as the intro content is fleshed out further.