Bump docs-sourcer version (#299)
This PR bumps the docs-sourcer version to fix the build issues that we were having. When you pull you should run `yarn install` in order to make sure you get the latest package versions.
Also updating all of the locally sourced content as the new versions of the docs-sourcer formats the metadata differently at the end of the file.