[APT-2107] Added canonical URL for KB pages (#585)
This PR adds generated KB pages that define a canonical URL that points back to the original GitHub discussion so that SEO is not impacted by duplicate content. It's very unlikely that SEO would be impacted even without this tag as the content of what is in GitHub and on our docs site is significantly different as we only include the original description and just the selected answer. We don't copy structure or pull in the other comments. Still - the change in this PR is a further step to ensure correctness.