Update Production Framework guide to remove roadmap items (#648)
* Update Production Framework guide to remove roadmap items

A customer reported confusion on a roadmap commitment we made here. We shouldn't be making roadmap commitments here since it's too hard to keep track of them and they'll inevitably get out of date.

So I've updated the "How Gruntwork Helps" column to reflect our latest offerings and removed the "On the Roadmap" column.