[CORE-569] Add "Sample Usage" section to modules
This PR adds a new section into the README of the reference documentation for Gruntwork's modules. The new section displays sample terraform source code with the module's source, require and optional variables. This example can be quickly copied to get started with using the module. Each variable lists the description as a comment on top of it.