[WIP] Global Docs Search
This PR seeks to serve as a base for our work in replacing the search on the docs site with a custom component that's capable of doing contextual search using multiple Algolia indices. It also adds Tailwind to the docs site.

This PR is not meant to be reviewed or merged at this time.